requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
